Basic energy smearing and Spectrum output
Implemented energy smearing of produced electrons based solely on the fano factor and a gaussian (poisson) dist. I have also rounded the energy values to integers, although this is unnecessary since the histogram that is outputting them spectrum would be able to bin the spectrum itself.

So far the channel number and detected energies are 1to1, as if the gain has been adjusted perfectly with accordance to calibrations.