Decay time
Implemented the probability of a photon interacting within the detector volume based off of the cross section and number density (for NaI(Tl)). Used this to also get an efficiency of the detector. In the future I may integrate over the lifetime rather than calculate it outright based on the total expected distance