Support for different energies
Allows the user to specify the energy of the source (needed to renormalise the probability distribution for this).

I have also noticed that there is a discontinuity in the spectra not explained by any physics, and it seems to happen at energy values that correspond to a 90degree scatter, so I will look into this.