GOOEY
Added a basic gui with the ability to change some parameters as a user. Resizing it is weird, I am still getting the hang of pyqt layout shenanigans.
Ideally I would want the batches readout to be displayed in the window but since it runs in the backend file it makes the entire window unresponsive while running. Maybe I'll combine the two files.
main -> main_backend
qtwidget -> main_frontend

The program must also be ran from main_frontend now due to circular referencing being an issue if attempted to run in main_backend (previously main)