Added 2d representation of detection
It is displayed in degrees rather than radians, so the Y axis ranges from -90 to +90, and the X axis ranges from 0 to 360.

However, this has brought to light an issue with the detection algorithm extending vectors backwards and creating a detection where there shouldn't be one. I need to figure out how to fix this. But I am glad that this has been brought to light since I've been wondering why the detection efficiencies have been so high, and this is affecting them due to how multiple scatters are detected.